snRuntime tests: reorganize startup files for rtl, banshee general, banshee snitch and make necessary cmake adaptations
