+++ sw/banshee/config/snitch_cluster.yaml
@@ -1,0 +1,46 @@
+# Copyright 2021 ETH Zurich and University of Bologna.
+# Licensed under the Apache License, Version 2.0, see LICENSE for details.
+# SPDX-License-Identifier: Apache-2.0
+
+---
+address:
+  barrier_reg: 0x40000038
+  cluster_base_hartid: 0x40000040
+  cluster_id: 0x40000050
+  cluster_num: 0x40000048
+  nr_cores: 0x40000010
+  scratch_reg: 0x40000020
+  tcdm_end: 0x40000008
+  tcdm_start: 0x40000000
+  uart: 0xF00B8000
+  wakeup_reg: 0x40000028
+  clint: 0xFFFF0000 # Snitch: Bootdata.clint_base
+  cl_clint: 0x120030 # Snitch: cl_peripherals + 0x30
+architecture:
+  num_cores: 9 # per cluster
+  num_clusters: 1
+  base_hartid: 0
+bootrom:
+  callbacks:
+    - name: bootrom-cluster
+      size: 0x34
+  end: 0x01020000
+  latency: 5
+  start: 0x01000000
+memory:
+- dram:
+    end: 0xF0000000
+    latency: 4
+    start: 0x80000000
+  ext_tcdm: []
+  periphs:
+    callbacks:
+    - name: zero-memory
+      size: 64000
+    end: 0x140000
+    latency: 2
+    start: 0x130000
+  tcdm:
+    end: 0x120000
+    latency: 1
+    start: 0x100000
