+++ sw/snRuntime/src/platforms/banshee-general/start_banshee.c
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 #include "snrt.h"
-#include "team.h"
+#include "../../team.h"
 
 extern const uint32_t _snrt_banshee_cluster_core_num;
 extern const uint32_t _snrt_banshee_cluster_base_hartid;
