+++ sw/snRuntime/src/platforms/banshee/putchar.c
@@ -1,0 +1,11 @@
+// Copyright 2020 ETH Zurich and University of Bologna.
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+#include "snrt.h"
+
+extern const uint32_t wakeup_reg;
+
+// Provide an implementation for putchar.
+void snrt_putchar(char character) {
+    *(volatile uint32_t *)0xF00B8000 = character;
+}
