+++ sw/snRuntime/src/platforms/shared/start.c
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 #include "snitch_cluster_peripheral.h"
 #include "snrt.h"
-#include "team.h"
+#include "../../team.h"
 
 extern const uint32_t _snrt_cluster_cluster_core_num;
 extern const uint32_t _snrt_cluster_cluster_base_hartid;
@@ -14,7 +14,7 @@ void *const _snrt_cluster_global_offset = (void *)0x10000000;
 const uint32_t snrt_stack_size __attribute__((weak, section(".rodata"))) = 10;
 
 // The boot data generated along with the system RTL.
-// See `hw/system/snitch_cluster/test/tb_lib.hh` for details.
+// See `ip/test/src/tb_lib.hh` for details.
 struct snrt_cluster_bootdata {
     uint32_t boot_addr;
     uint32_t core_count;
@@ -100,25 +100,4 @@ void _snrt_init_team(uint32_t cluster_core_id, uint32_t cluster_core_num,
 
 uint32_t _snrt_barrier_reg_ptr() {
     return _snrt_team_current->root->barrier_reg_ptr;
-}
-
-extern uintptr_t volatile tohost, fromhost;
-
-// Provide an implementation for putchar.
-void snrt_putchar(char character) {
-    volatile struct putc_buffer *buf = &putc_buffer[snrt_hartid()];
-    buf->data[buf->hdr.size++] = character;
-    if (buf->hdr.size == PUTC_BUFFER_LEN || character == '\n') {
-        buf->hdr.syscall_mem[0] = 64;  // sys_write
-        buf->hdr.syscall_mem[1] = 1;   // file descriptor (1 = stdout)
-        buf->hdr.syscall_mem[2] = (uintptr_t)&buf->data;  // buffer
-        buf->hdr.syscall_mem[3] = buf->hdr.size;          // length
-
-        tohost = (uintptr_t)buf->hdr.syscall_mem;
-        while (fromhost == 0)
-            ;
-        fromhost = 0;
-
-        buf->hdr.size = 0;
-    }
 }
