+++ sw/snRuntime/src/platforms/shared/start_snitch.S
@@ -2,7 +2,7 @@
 # Licensed under the Apache License, Version 2.0, see LICENSE for details.
 # SPDX-License-Identifier: Apache-2.0
 
-#include "start.S"
+#include "../../start.S"
 
 # Function to determine the initial cluster layout to setup the initial stack.
 # See `hw/system/snitch_cluster/test/tb_lib.hh` for details of the bootdata
@@ -27,18 +27,6 @@ _snrt_init_core_info:
     remu      a0, a0, a1 # calculate cluster-local core id
     ret
 
-# Must leave a0 untouched.
-_snrt_cluster_barrier:
-    .globl _snrt_cluster_barrier
-    addi      sp, sp, -4
-    sw        ra, 0(sp)
-    call      _snrt_barrier_reg_ptr
-    lw        a0, 0(a0)
-    mv        zero, a0
-    lw        ra, 0(sp)
-    addi      sp, sp, 4
-    ret
-
 # Mustn't use any non-temporary registers
 # Expecting:
 # - a0: device tree ptr ("bootdata" in `tb_lib.hh`)
@@ -47,32 +35,3 @@ _snrt_get_base_hartid:
     .globl _snrt_get_base_hartid
     lw        a0, 8(a0)  # load cluster's hartid offset
     ret
-
-# Function to terminate execution.
-#
-# Expecting:
-# - a0: exit code
-# - a1: device tree ptr
-_snrt_exit:
-    .globl _snrt_exit
-    # Determine global core index.
-    addi      sp, sp, -8
-    sw        a0, 0(sp)
-    sw        ra, 4(sp)
-    call      snrt_global_core_idx
-    lw        t0, 0(sp)
-    lw        ra, 4(sp)
-    addi      sp, sp, 8
-
-    # Only first core triggers exit.
-    bnez      a0, 1f
-    slli      t0, t0, 1
-    ori       t0, t0, 1
-    la        t1, tohost
-    sw        t0, 0(t1)
-1:  ret
-
-# HTIF sections
-.pushsection .htif,"aw",@progbits;
-.align 6; .global tohost; tohost: .dword 0;
-.align 6; .global fromhost; fromhost: .dword 0;
